% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction_factors.R
\docType{data}
\name{correction_factors}
\alias{correction_factors}
\title{Quadrupole bias correction factors}
\format{
A tibble with 62 rows and
3 variables:

\describe{
\item{metabolite}{name of metabolite}
\item{M}{specific isotope}
\item{cf}{correction factor that is multiplied by peak area}
}
}
\usage{
correction_factors
}
\description{
A dataset containing the experimentally determined peak area correction
factors to adjust for differential isotope transit through the quadrupole.
}
\keyword{datasets}
