% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negate \%in\%}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{The values to be matched}

\item{y}{The values to be matched against}
}
\description{
Binary operator that returns a logical vector returning \code{TRUE} if there is
not a match for its left operand.
}
