% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb_mappings.R
\docType{data}
\name{hmdb_mappings}
\alias{hmdb_mappings}
\title{HMDB IDs for targeted metabolomics metabolite names}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 141 rows and 2 columns.
}
\usage{
hmdb_mappings
}
\description{
\describe{
\item{metabolite}{name of metabolite}
\item{hmdb}{HMDB ID}
}
}
\keyword{datasets}
