% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{make_std_plots}
\alias{make_std_plots}
\title{Generate generic standard curves}
\usage{
make_std_plots(df, title = NULL)
}
\arguments{
\item{df}{A tibble containing data for a standard curve}

\item{title}{Optional to label the standard curve}
}
\value{
A ggplot
}
\description{
Generate generic standard curves
}
