% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Add names to tibble list produced from group_split}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{A tibble}

\item{...}{Variables to group by}
}
\value{
A named list of tibbles.
}
\description{
Add names to tibble list produced from group_split
}
