% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\docType{data}
\name{ppi}
\alias{ppi}
\title{Protein-Protein Interactome}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 16470.
}
\usage{
ppi
}
\description{
The consolidated human protein-protein interactome. Contains 233,957 edges
between 16,470 nodes. Current as of June 2018.
}
\keyword{datasets}
