% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_plots}
\alias{print_plots}
\title{Save multiple ggplots}
\usage{
print_plots(plots, names, path, width = 20, height = 15)
}
\arguments{
\item{plots}{List of plots to save}

\item{names}{Base filenames of plots}

\item{path}{Directory to save plots}

\item{width}{Plot width in cm}

\item{height}{Plot height in cm}
}
\description{
Save multiple ggplots
}
