% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_multi_excel}
\alias{read_multi_excel}
\title{Reads and combines multi-sheet Excel files}
\usage{
read_multi_excel(excel_file_list)
}
\arguments{
\item{excel_file_list}{A preferably named list of Excel files.}
}
\value{
A list of tibbles for each unique Excel sheet in the submitted files.
}
\description{
This function extracts and combines Excel sheets from multiple files.
}
