% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers using median absolute deviation}
\usage{
remove_outliers(tbl_df, column, remove = FALSE)
}
\arguments{
\item{tbl_df}{Tibble containing data to be filtered}

\item{column}{Column name of tibble from which outliers are to be removed}

\item{remove}{If TRUE, removes outlier observations from returned tibble}
}
\value{
If remove = FALSE, a tibble containing a new column, \code{filtered},
where filtered values have been replaced by NA. If remove = TRUE, outlier
values have been removed from the tibble.
}
\description{
This function takes a tibble and will remove outliers from a specific column
using a median absolute deviation > 2. It returns the original tibble with a
new column \code{filtered} that contains NA in place of filtered values.
}
\examples{
remove_outliers(ggplot2::diamonds, "price")
}
