% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmo-package.R
\docType{package}
\name{wmo-package}
\alias{wmo}
\alias{wmo-package}
\title{wmo: Personal Utility Functions}
\description{
Makes commonly used utility functions, data sets, and document
    templates available in a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/wmo}
  \item Report bugs at \url{https://github.com/wmoldham/wmo/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
