% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate x values from a standard curve}
\usage{
interpolate(new_df, model)
}
\arguments{
\item{new_df}{Data frame containing new y values}

\item{model}{Linear model}
}
\value{
List of interpolated x values
}
\description{
This function takes a model object and a data frame containing new y values
and will interpolate the associated x values. This is useful for determining
sample concentrations from a standard curve. The intention is that it will
operate similarly to \code{predict()}. Currently, it will support polynomial fits.
}
\details{
There is nothing yet in the function that protects from multiple roots being
returned. In addition, it will return \code{NA} if the new x values lie outside of
the range of the x values in the model.
}
\examples{
x <- seq(from = 2, to = 10, by = 2)
y <- 2 * x
m <- lm(y ~ x)
new_y <- data.frame(y = seq(from = 1, to = 11, by = 2))
interpolate(new_y, m)
}
